# ⚡️ Speed up `describe()` by 6x