--- a/docs/arrays.ipynb
+++ b/docs/arrays.ipynb
@@ -126,7 +126,7 @@
    },
    "outputs": [],
    "source": [
-    "ln.Artifact(\"hf://datasets/Koncopd/lamindb-test/sharded_parquet\").save()"
+    "# ln.Artifact(\"hf://datasets/Koncopd/lamindb-test/sharded_parquet\").save()"
    ]
   },
   {
@@ -146,16 +146,16 @@
    },
    "outputs": [],
    "source": [
-    "artifact_shard1 = ln.Artifact(\n",
-    "    \"hf://datasets/Koncopd/lamindb-test/sharded_parquet/louvain=0/947eee0b064440c9b9910ca2eb89e608-0.parquet\"\n",
-    ").save()\n",
-    "artifact_shard2 = ln.Artifact(\n",
-    "    \"hf://datasets/Koncopd/lamindb-test/sharded_parquet/louvain=1/947eee0b064440c9b9910ca2eb89e608-0.parquet\"\n",
-    ").save()\n",
+    "# artifact_shard1 = ln.Artifact(\n",
+    "#     \"hf://datasets/Koncopd/lamindb-test/sharded_parquet/louvain=0/947eee0b064440c9b9910ca2eb89e608-0.parquet\"\n",
+    "# ).save()\n",
+    "# artifact_shard2 = ln.Artifact(\n",
+    "#     \"hf://datasets/Koncopd/lamindb-test/sharded_parquet/louvain=1/947eee0b064440c9b9910ca2eb89e608-0.parquet\"\n",
+    "# ).save()\n",
     "\n",
-    "ln.Collection(\n",
-    "    [artifact_shard1, artifact_shard2], key=\"sharded_parquet_collection\"\n",
-    ").save()"
+    "# ln.Collection(\n",
+    "#     [artifact_shard1, artifact_shard2], key=\"sharded_parquet_collection\"\n",
+    "# ).save()"
    ]
   },
   {
@@ -395,7 +395,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "artifact = ln.Artifact.get(key=\"sharded_parquet\")"
+    "# artifact = ln.Artifact.get(key=\"sharded_parquet\")"
    ]
   },
   {
@@ -408,7 +408,7 @@
    },
    "outputs": [],
    "source": [
-    "artifact.path.view_tree()"
+    "# artifact.path.view_tree()"
    ]
   },
   {
@@ -421,7 +421,7 @@
    },
    "outputs": [],
    "source": [
-    "backed = artifact.open()"
+    "# backed = artifact.open()"
    ]
   },
   {
@@ -437,7 +437,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "backed"
+    "# backed"
    ]
   },
   {
@@ -450,7 +450,7 @@
    },
    "outputs": [],
    "source": [
-    "backed.head(5).to_pandas()"
+    "# backed.head(5).to_pandas()"
    ]
   },
   {
@@ -466,7 +466,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "collection = ln.Collection.get(key=\"sharded_parquet_collection\")"
+    "# collection = ln.Collection.get(key=\"sharded_parquet_collection\")"
    ]
   },
   {
@@ -475,7 +475,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "backed = collection.open()"
+    "# backed = collection.open()"
    ]
   },
   {
@@ -484,7 +484,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "backed"
+    "# backed"
    ]
   },
   {
@@ -497,7 +497,7 @@
    },
    "outputs": [],
    "source": [
-    "backed.to_table().to_pandas()"
+    "# backed.to_table().to_pandas()"
    ]
   },
   {
@@ -517,8 +517,8 @@
    },
    "outputs": [],
    "source": [
-    "with collection.open(engine=\"polars\") as lazy_df:\n",
-    "    display(lazy_df.collect().to_pandas())"
+    "# with collection.open(engine=\"polars\") as lazy_df:\n",
+    "#     display(lazy_df.collect().to_pandas())"
    ]
   },
   {
@@ -534,7 +534,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "backed = ln.Artifact.filter(suffix=\".parquet\").open()"
+    "# backed = ln.Artifact.filter(suffix=\".parquet\").open()"
    ]
   },
   {
@@ -543,7 +543,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "backed"
+    "# backed"
    ]
   },
   {
@@ -551,7 +551,7 @@
    "execution_count": null,
    "metadata": {
     "tags": [
-     "hide-cell"
+     "hide-output"
     ]
    },
    "outputs": [],