# Summary

Testing for a rather heavily annotated artifact on the other side of the Atlantic.

```
LINE_PROFILE=1 lamin describe --key example_datasets/mini_immuno/dataset1.h5ad
```

Before | After
--- | ---
8.1 sec | 1.3 sec
<img width="895" height="604" alt="image" src="https://github.com/user-attachments/assets/c4e49030-e40f-44d1-9925-1a49d087b367" /> | <img width="1054" height="652" alt="image" src="https://github.com/user-attachments/assets/0f8ef12e-563e-490a-a063-a7df74e04596" />

# Steps

## No more storage access: reduce by 2.2 sec

https://github.com/laminlabs/lamindb/pull/2953/commits/32fac74915be857ae71dd4f02385e6477487f86c

Before | After
--- | ---
2.3 sec | 0 sec
<img width="787" height="198" alt="image" src="https://github.com/user-attachments/assets/ed054f63-bc29-4eb9-9e11-9e44f622c61b" /> | <img width="719" height="225" alt="image" src="https://github.com/user-attachments/assets/a4a38f22-dd1b-4c11-89da-235268aa2584" />

## No longer treating links and M2M in two queries: reduce by 4 sec

https://github.com/laminlabs/lamindb/pull/2953/commits/82c5f8d5e8380199b67ae9d43e3a1e23989b4535

Before | After
--- | ---
4 sec (93% of pre-fetching) | 0 sec
<img width="729" height="236" alt="image" src="https://github.com/user-attachments/assets/67731847-0895-4897-8c73-4b1271a538bc" /> | <img width="700" height="227" alt="image" src="https://github.com/user-attachments/assets/f352a804-2581-4ad7-8052-feda3573d165" />
<img width="962" height="659" alt="image" src="https://github.com/user-attachments/assets/06f6b083-4ca4-48d7-9bf1-50edf6cd7d08" /> | <img width="1199" height="341" alt="image" src="https://github.com/user-attachments/assets/67ec4399-f6ae-4a5a-95c8-2ff6444d7d68" />

## Leverage pre-fetched foreign keys: reduction of 0.66 sec

https://github.com/laminlabs/lamindb/pull/2953/commits/a8943173607ae9f51cafc21e27a9d64ca36f3db5

Before | After
--- | ---
0.66 sec | 0 sec
<img width="942" height="284" alt="image" src="https://github.com/user-attachments/assets/8dadf679-d7c2-45e2-a98a-4c218033e1c2" /> | <img width="1367" height="179" alt="image" src="https://github.com/user-attachments/assets/afa1bdb2-0f18-42f0-b535-b5134c91c641" />